+++ controllers/api_controllers/getAPI.js
@@ -1,4 +1,4 @@
-const endpoints = require("../endpoints.json")
+const endpoints = require("../../endpoints.json")
 
 exports.getAPI = (req,res)=>{
     res.status(200).send({endpoints})
