--- controllers/getCommentsByArticleId.js
@@ -1,17 +1,0 @@
-const { fetchArticleById } = require("../models/fetchArticleById");
-const { fetchCommentsByArticleId } = require("../models/fetchCommentsByArticleId");
-
-exports.getCommentsByArticleId = (req, res, next) => {
-  const { article_id } = req.params;
-  const promises = [fetchArticleById(article_id)];
-  promises.push(fetchCommentsByArticleId(article_id));
-
-  return Promise.all(promises)
-    .then(([_, comments]) => {
-      if (!comments.length) {
-        res.status(404).send({ msg: `No comments found for article_id: ${article_id}` });
-      }
-      res.status(200).send({ comments });
-    })
-    .catch(next);
-};
