+++ models/topic_models/fetchAllTopics.js
@@ -1,8 +1,7 @@
-const db = require("../db/connection")
+const db = require("../../db/connection")
 
 exports.fetchAllTopics = ()=>{
     return db.query(`SELECT * FROM topics`).then(({rows})=>{
-        console.log(rows)
         return rows
     })
 }
